Can now add evaluation to a whole class by selected subject - can also display evaluations for a class JS selecting subject

+++ db/migrate/20120623121528_create_evaluations.rb
@@ -4,7 +4,7 @@ class CreateEvaluations < ActiveRecord::Migration
       t.integer :student_id
       t.integer :subject_id
       t.float :grade
-      t.float :scale
+      t.integer :scale
 
       t.timestamps
     end
